requirements.txt
Django>=4.2.0
channels>=4.0.0
channels-redis>=4.1.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
Pillow>=9.5.0
python-dotenv>=1.0.0
mysqlclient>=2.1.1